+++ src/api/layers/retriever.layer.ts
@@ -180,32 +180,14 @@ export class RetrieverLayer extends SenderLayer {
    * Checks if a number is a valid WA number
    * @category Contact
    * @param contactId, you need to include the @c.us at the end.
-   * @returns contact detial as promise
+   * @returns contact details as promise
    */
   public async checkNumberStatus(contactId: string): Promise<WhatsappProfile> {
-    const result = await evaluateAndReturn(
+    return await evaluateAndReturn(
       this.page,
-      (contactId) => WPP.contact.queryExists(contactId),
+      (contactId) => WAPI.checkNumberStatus(contactId),
       contactId
     );
-
-    if (!result) {
-      return {
-        id: contactId as any,
-        isBusiness: false,
-        canReceiveMessage: false,
-        numberExists: false,
-        status: 404,
-      };
-    }
-
-    return {
-      id: result.wid as any,
-      isBusiness: result.biz,
-      canReceiveMessage: true,
-      numberExists: true,
-      status: 200,
-    };
   }
 
   /**
@@ -274,7 +256,7 @@ export class RetrieverLayer extends SenderLayer {
    * Retrieves contact detail object of given contact id
    * @category Contact
    * @param contactId
-   * @returns contact detial as promise
+   * @returns contact details as promise
    */
   public async getContact(contactId: string) {
     return evaluateAndReturn(
@@ -297,9 +279,9 @@ export class RetrieverLayer extends SenderLayer {
    * Retrieves chat object of given contact id
    * @category Chat
    * @param contactId
-   * @returns contact detial as promise
+   * @returns chat details as promise
    */
-  public async getChatById(contactId: string): Promise<Chat> {
+  public async getChatById(contactId: string | Wid): Promise<Chat> {
     return evaluateAndReturn(
       this.page,
       (contactId) => WAPI.getChatById(contactId),
@@ -311,10 +293,10 @@ export class RetrieverLayer extends SenderLayer {
    * Retrieves chat object of given contact id
    * @category Chat
    * @param contactId
-   * @returns contact detial as promise
+   * @returns chat details as promise
    * @deprecated
    */
-  public async getChat(contactId: string) {
+  public async getChat(contactId: string | Wid) {
     return this.getChatById(contactId);
   }
 
@@ -341,7 +323,7 @@ export class RetrieverLayer extends SenderLayer {
    * @deprecated Depreciado em favor de getMessages
    * @category Chat
    * @param contactId
-   * @returns contact detial as promise
+   * @returns contact details as promise
    */
   public async loadEarlierMessages(contactId: string) {
     return evaluateAndReturn(
@@ -378,7 +360,7 @@ export class RetrieverLayer extends SenderLayer {
    * @deprecated Deprecated in favor of checkNumberStatus
    * @category Contact
    * @param contactId, you need to include the @c.us at the end.
-   * @returns contact detial as promise
+   * @returns contact details as promise
    */
   public async getNumberProfile(contactId: string) {
     this.log(
@@ -631,6 +613,27 @@ export class RetrieverLayer extends SenderLayer {
       this.page,
       ({ wid }) => WPP.contact.getCommonGroups(wid),
       { wid }
+    );
+  }
+
+  /**
+   * Get LID/PhoneNumber mapping and Contact information
+   *
+   * @example
+   * ```javascript
+   * const info = await client.getPnLidEntry('[number]@c.us');
+   * const info = await client.getPnLidEntry('[number]@lid');
+   * ```
+   *
+   * @category Contact
+   * @param phoneOrLid Contact ID (phone number or LID)
+   * @returns Promise with lid, phoneNumber and contact information
+   */
+  public async getPnLidEntry(phoneOrLid: string) {
+    return await evaluateAndReturn(
+      this.page,
+      (phoneOrLid) => WPP.contact.getPnLidEntry(phoneOrLid),
+      phoneOrLid
     );
   }
 }