+++ src/api/layers/ui.layer.ts
@@ -18,6 +18,7 @@
 import { Page } from 'puppeteer';
 import { CreateConfig } from '../../config/create-config';
 import { evaluateAndReturn } from '../helpers';
+import { Wid, Chat } from '../model';
 import { GroupLayer } from './group.layer';
 
 export class UILayer extends GroupLayer {
@@ -31,10 +32,10 @@ export class UILayer extends GroupLayer {
    * @category UI
    * @param chatId
    */
-  public async openChat(chatId: string) {
+  public async openChat(chatId: string | Wid) {
     return evaluateAndReturn(
       this.page,
-      (chatId: string) => WPP.chat.openChatBottom(chatId),
+      (chatId: string) => WPP.chat.openChatBottom(chatId, undefined),
       chatId
     );
   }
@@ -45,18 +46,30 @@ export class UILayer extends GroupLayer {
    * @param chatId Chat id
    * @param messageId Message id (For example: '06D3AB3D0EEB9D077A3F9A3EFF4DD030')
    */
-  public async openChatAt(chatId: string, messageId: string) {
+  public async openChatAt(chatId: string | Wid, messageId: string) {
     return evaluateAndReturn(
       this.page,
-      (chatId: string) => WPP.chat.openChatAt(chatId, messageId),
-      chatId
+      (chatId: string, messageId) =>
+        WPP.chat.openChatAt(chatId, messageId, undefined),
+      chatId,
+      messageId
     );
   }
+
   /**
-   * Return the current active chat
+   * Closes the currently opened chat (if any).
+   * The boolean result reflects if there was any chat that got closed.
    * @category UI
    */
-  public async getActiveChat() {
+  public async closeChat() {
+    return evaluateAndReturn(this.page, () => WPP.chat.closeChat());
+  }
+
+  /**
+   * Return the currently active chat (visually open)
+   * @category UI
+   */
+  public getActiveChat() {
     return evaluateAndReturn(this.page, () => WPP.chat.getActiveChat());
   }
 }